+++ react/src/js/components/Consonant/Container/__tests__/LeftFilter.spec.js
@@ -108,7 +108,6 @@ describe('Consonant/Container/Left Filter', () => {
         expect(filtersLeftElement).toHaveClass('is-opened');
     });
 
-
     test('should be able to search and filter', async () => {
         const configToUse = config;
         configToUse.filterPanel.filterLogic = 'xor';